refactor(plines): move out plines related code from misc1.c
