+++ src/nvim/fold.c
@@ -29,6 +29,7 @@
 #include "nvim/memory.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/plines.h"
 #include "nvim/garray.h"
 #include "nvim/move.h"
 #include "nvim/option.h"
